+++ paillmtrace-20240311/ChangeLog.md
@@ -1,3 +1,10 @@
+2025-08-01 Version: 1.1.1
+- Update API ListOnlineEvalTasks: add request parameters SortBy.
+- Update API ListOnlineEvalTasks: add request parameters SortOrder.
+- Update API ListOnlineEvalTasks: add response parameters Body.Tasks.$.EvalResults.
+- Update API ListOnlineEvalTasks: add response parameters Body.Tasks.$.RecordCount.
+
+
 2025-07-08 Version: 1.1.0
 - Support API ListEvalResults.
 - Update API ListTracesDatas: add request parameters MaxDuration.